+++ src/Deadliner/DeadlinesDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="AddDeadlineDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="Deadliner.DeadlinesDialog">
   <grid id="cbd77" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1" binding="contentPane">
     <margin top="10" left="10" bottom="10" right="10"/>
     <constraints>
