+++ src/Deadliner/SubjectButton.java
@@ -1,3 +1,5 @@
+package Deadliner;
+
 import javax.swing.*;
 import java.awt.*;
