+++ src/Deadliner/Week.java
@@ -1,5 +1,8 @@
+package Deadliner;
+
+import Deadliner.Class;
+
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.util.Scanner;
